app/requirements.txt
fastapi==0.95.1
pydantic==1.10.7
uvicorn==0.21.1
python-jose==3.3.0 
passlib==1.7.4
SQLAlchemy==2.0.10
PyJWT==2.6.0
psycopg2-binary==2.9.6
bcrypt==4.0.1
python-multipart
pydantic[email]